# contracts/ts/src/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: `${string}://contracts` | (string & {});
};

/**
 * AuthenticatedPrincipalResponse
 */
export type AuthenticatedPrincipalResponse = {
    /**
     * Issuer
     */
    issuer: string;
    /**
     * Scopes
     */
    scopes: Array<string>;
    /**
     * Subject
     */
    subject: string;
};

/**
 * ProblemResponse
 */
export type ProblemResponse = {
    /**
     * Code
     */
    code: string;
    /**
     * Detail
     */
    detail: string;
    /**
     * Request Id
     */
    request_id: string;
    /**
     * Retryable
     */
    retryable: boolean;
    /**
     * Status
     */
    status: number;
    /**
     * Title
     */
    title: string;
    /**
     * Type
     */
    type: string;
};

/**
 * ServiceStatusResponse
 */
export type ServiceStatusResponse = {
    /**
     * Api Version
     */
    api_version: 'v1';
    /**
     * Service
     */
    service: 'flash-trips-api';
    /**
     * Status
     */
    status: 'ok';
};

export type GetAuthenticatedPrincipalData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/v1/authenticated-principal';
};

export type GetAuthenticatedPrincipalErrors = {
    /**
     * Unauthorized
     */
    401: ProblemResponse;
    /**
     * Forbidden
     */
    403: ProblemResponse;
    /**
     * Not Found
     */
    404: ProblemResponse;
    /**
     * Internal Server Error
     */
    500: ProblemResponse;
};

export type GetAuthenticatedPrincipalError = GetAuthenticatedPrincipalErrors[keyof GetAuthenticatedPrincipalErrors];

export type GetAuthenticatedPrincipalResponses = {
    /**
     * Successful Response
     */
    200: AuthenticatedPrincipalResponse;
};

export type GetAuthenticatedPrincipalResponse = GetAuthenticatedPrincipalResponses[keyof GetAuthenticatedPrincipalResponses];

export type GetServiceStatusData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/v1/status';
};

export type GetServiceStatusErrors = {
    /**
     * Not Found
     */
    404: ProblemResponse;
    /**
     * Internal Server Error
     */
    500: ProblemResponse;
};

export type GetServiceStatusError = GetServiceStatusErrors[keyof GetServiceStatusErrors];

export type GetServiceStatusResponses = {
    /**
     * Successful Response
     */
    200: ServiceStatusResponse;
};

export type GetServiceStatusResponse = GetServiceStatusResponses[keyof GetServiceStatusResponses];