# src/flash_trips/adapters/http/problems.py
from collections.abc import Mapping
from uuid import UUID

from fastapi import Request
from pydantic import BaseModel, ConfigDict
from starlette.exceptions import HTTPException
from starlette.middleware.base import BaseHTTPMiddleware, RequestResponseEndpoint
from starlette.responses import JSONResponse, Response

from flash_trips.kernel.identifiers import uuid7
from flash_trips.kernel.problem import Problem


class ProblemResponse(BaseModel):
    model_config = ConfigDict(frozen=True)

    type: str
    title: str
    status: int
    detail: str
    code: str
    retryable: bool
    request_id: UUID


class RequestIdMiddleware(BaseHTTPMiddleware):
    async def dispatch(
        self, request: Request, call_next: RequestResponseEndpoint
    ) -> Response:
        request_id = str(uuid7())
        request.state.request_id = request_id
        response = await call_next(request)
        response.headers["X-Request-ID"] = str(request_id)
        return response


async def http_problem(request: Request, error: Exception) -> JSONResponse:
    if not isinstance(error, HTTPException):
        raise error

    if error.status_code == 401:
        problem = Problem(
            type="https://flash-trips.example/problems/authentication-required",
            title="Unauthorized",
            status=401,
            detail="Authentication is required.",
            code="authentication_required",
            retryable=False,
            request_id=request.state.request_id,
        )
    elif error.status_code == 403:
        problem = Problem(
            type="https://flash-trips.example/problems/access-denied",
            title="Forbidden",
            status=403,
            detail="Access is denied.",
            code="access_denied",
            retryable=False,
            request_id=request.state.request_id,
        )
    elif error.status_code == 404:
        problem = Problem(
            type="https://flash-trips.example/problems/route-not-found",
            title="Not Found",
            status=404,
            detail="The requested resource was not found.",
            code="route_not_found",
            retryable=False,
            request_id=request.state.request_id,
        )
    else:
        problem = Problem(
            type="https://flash-trips.example/problems/http-error",
            title="Request Failed",
            status=error.status_code,
            detail="The request could not be completed.",
            code="http_error",
            retryable=False,
            request_id=request.state.request_id,
        )

    return problem_json(problem, headers=error.headers)


async def unhandled_problem(request: Request, error: Exception) -> JSONResponse:
    del error
    return problem_json(
        Problem(
            type="https://flash-trips.example/problems/internal-error",
            title="Internal Server Error",
            status=500,
            detail="The request could not be completed.",
            code="internal_error",
            retryable=True,
            request_id=request.state.request_id,
        )
    )


def problem_json(
    problem: Problem,
    *,
    headers: Mapping[str, str] | None = None,
) -> JSONResponse:
    request_id = str(problem.request_id)
    response_headers = dict(headers or {})
    response_headers["X-Request-ID"] = request_id
    return JSONResponse(
        ProblemResponse.model_validate(problem, from_attributes=True).model_dump(
            mode="json"
        ),
        status_code=problem.status,
        media_type="application/problem+json",
        headers=response_headers,
    )